telemetry-python/minirequirements.txt
argparse-addons==0.12.0
bitstruct==8.17.0
cantools==38.0.2
crccheck==1.3.0
digi-xbee==1.4.1
diskcache==5.4.0
msgpack==1.0.4
packaging==23.0
pyserial==3.5
python-can==4.1.0
textparser==0.24.0
typing_extensions==4.5.0
wrapt==1.14.1
